Add safety and quiet mode.
Previously, indexes were blindly deleted before population.  Because you
might not want to blow away an index, it will ask before doing so now
unless the unsafeIndexDelete flag is set to true in the config.

Also, added a quiet mode to suppress the success output.

Finally, it'd be nice to be able to not delete the index at all and
instead increment to an existing one.